+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/internal/AbstractIgnoredAnnotationRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.rule;
+package net.sourceforge.pmd.lang.java.rule.internal;
 
 import static net.sourceforge.pmd.properties.PropertyFactory.stringListProperty;
 
@@ -10,15 +10,13 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.java.ast.Annotatable;
+import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 
 /**
- * @deprecated Internal API
+ * @apiNote  Internal API
  */
-@Deprecated
-@InternalApi
 public abstract class AbstractIgnoredAnnotationRule extends AbstractJavaRule {
 
     private final PropertyDescriptor<List<String>> ignoredAnnotationsDescriptor